+++ apps/extension/src/ui/apps/dashboard/routes/Settings/Accounts/TreeItem.tsx
@@ -1,12 +1,26 @@
-import { UniqueIdentifier } from "@dnd-kit/core"
+import { AccountJsonAny } from "@core/domains/accounts/types"
+import { DraggableAttributes } from "@dnd-kit/core"
+import { SyntheticListenerMap } from "@dnd-kit/core/dist/hooks/utilities"
+import { WithTooltip } from "@talisman/components/Tooltip"
+import {
+  ChevronDownIcon,
+  DragIcon,
+  FolderIcon,
+  MoreHorizontalIcon,
+  TrashIcon,
+} from "@talisman/theme/icons"
 import { classNames } from "@talismn/util"
-import { HTMLAttributes, forwardRef } from "react"
+import { AccountIcon } from "@ui/domains/Account/AccountIcon"
+import { Address } from "@ui/domains/Account/Address"
+import Fiat from "@ui/domains/Asset/Fiat"
+import { CSSProperties, ReactNode, forwardRef, useMemo } from "react"
 
+import { AccountsLogoStack } from "./AccountsLogoStack"
 import { Action } from "./Action"
-import { Handle } from "./Handle"
-import styles from "./TreeItem.module.css"
+import { UiTreeAccount, UiTreeFolder, UiTreeItem } from "./types"
 
-export interface Props extends Omit<HTMLAttributes<HTMLLIElement>, "id"> {
+export interface Props {
+  accounts: AccountJsonAny[]
   childCount?: number
   clone?: boolean
   collapsed?: boolean
@@ -14,73 +28,152 @@ export interface Props extends Omit<HTMLAttributes<HTMLLIElement>, "id"> {
   disableInteraction?: boolean
   disableSelection?: boolean
   ghost?: boolean
-  handleProps?: any
+  handleProps?: { attributes: DraggableAttributes; listeners?: SyntheticListenerMap }
   indentationWidth: number
-  value: UniqueIdentifier
+  item: UiTreeItem
+  style: CSSProperties
   onCollapse?(): void
+  onDelete?(): void
   wrapperRef?(node: HTMLLIElement): void
 }
 
-export const TreeItem = forwardRef<HTMLDivElement, Props>(
-  (
-    {
-      childCount,
-      clone,
-      depth,
-      disableSelection,
-      disableInteraction,
-      ghost,
-      handleProps,
-      indentationWidth,
-      collapsed,
-      onCollapse,
-      style,
-      value,
-      wrapperRef,
-      ...props
-    },
-    ref
-  ) => {
+export const TreeItem = forwardRef<HTMLDivElement, Props>((props, ref) => {
+  const { item } = props
+
+  if (item.type === "account") return <TreeAccountItem {...props} item={item} ref={ref} />
+  return <TreeFolderItem {...props} item={item} ref={ref} />
+})
+TreeItem.displayName = "TreeItem"
+
+export const TreeAccountItem = forwardRef<HTMLDivElement, Props & { item: UiTreeAccount }>(
+  (props, ref) => {
+    const { item, handleProps, style, accounts, wrapperRef, depth } = props
+    const account = accounts.find((account) => account.address === item.address)
+    if (!account) return null
+
     return (
-      <li
-        className={classNames(
-          styles.Wrapper,
-          clone && styles.clone,
-          ghost && styles.ghost,
-          disableSelection && styles.disableSelection,
-          disableInteraction && styles.disableInteraction
-        )}
-        ref={wrapperRef}
-        style={
-          {
-            "--spacing": `${indentationWidth * depth}px`,
-          } as React.CSSProperties
-        }
-        {...props}
-      >
-        <div className={styles.TreeItem} ref={ref} style={style}>
-          <Handle {...handleProps} />
-          {onCollapse && (
-            <Action
-              onClick={onCollapse}
-              className={classNames(styles.Collapse, collapsed && styles.collapsed)}
-            >
-              {collapseIcon}
-            </Action>
+      <TreeItemWrapper {...props} ref={wrapperRef}>
+        <div
+          ref={ref}
+          className={classNames(
+            "bg-black-secondary relative flex items-center gap-8 rounded-sm border-[1px] border-transparent p-8",
+            depth > 0 && "bg-black-secondary/60"
           )}
-          <span className={styles.Text}>{value}</span>
-          {clone && childCount && childCount > 1 ? (
-            <span className={styles.Count}>{childCount}</span>
-          ) : null}
+          style={style}
+        >
+          <DragButton {...handleProps?.attributes} {...handleProps?.listeners} />
+          <AccountIcon className="text-xl" address={item.address} />
+          <div className="flex grow flex-col gap-2 overflow-hidden">
+            <div className="overflow-hidden text-ellipsis whitespace-nowrap">{account.name}</div>
+            <div className="text-body-secondary text-sm">
+              <Address address={item.address} />
+            </div>
+          </div>
+          <div className="flex flex-col gap-2">
+            <WithTooltip as="div" tooltip="TODO: Add balance details" noWrap>
+              <Fiat amount={0} currency="usd" />
+            </WithTooltip>
+          </div>
+          <MoreHorizontalIcon className="shrink-0" />
         </div>
-      </li>
+      </TreeItemWrapper>
     )
   }
 )
-TreeItem.displayName = "TreeItem"
+TreeAccountItem.displayName = "TreeAccountItem"
 
-const collapseIcon = (
-  <svg width="10" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 70 41">
-    <path d="M30.76 39.2402C31.885 40.3638 33.41 40.995 35 40.995C36.59 40.995 38.115 40.3638 39.24 39.2402L68.24 10.2402C69.2998 9.10284 69.8768 7.59846 69.8494 6.04406C69.822 4.48965 69.1923 3.00657 68.093 1.90726C66.9937 0.807959 65.5106 0.178263 63.9562 0.150837C62.4018 0.123411 60.8974 0.700397 59.76 1.76024L35 26.5102L10.24 1.76024C9.10259 0.700397 7.59822 0.123411 6.04381 0.150837C4.4894 0.178263 3.00632 0.807959 1.90702 1.90726C0.807714 3.00657 0.178019 4.48965 0.150593 6.04406C0.123167 7.59846 0.700153 9.10284 1.75999 10.2402L30.76 39.2402Z" />
-  </svg>
+export const TreeFolderItem = forwardRef<HTMLDivElement, Props & { item: UiTreeFolder }>(
+  (props, ref) => {
+    const {
+      childCount,
+      clone,
+      handleProps,
+      item,
+      collapsed,
+      onCollapse,
+      onDelete,
+      style,
+      wrapperRef,
+    } = props
+
+    const addresses = useMemo(() => item.tree.map((item) => item.address), [item])
+
+    return (
+      <TreeItemWrapper {...props} ref={wrapperRef}>
+        {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events */}
+        <div
+          ref={ref}
+          role="button"
+          tabIndex={-1}
+          className="bg-black-secondary relative flex items-center gap-8 rounded-sm border-[1px] border-transparent p-8"
+          style={style}
+          onClick={onCollapse}
+        >
+          <DragButton {...handleProps?.attributes} {...handleProps?.listeners} />
+          <div className="bg-black-tertiary p-4 text-base">
+            <FolderIcon style={{ color: item.color }} />
+          </div>
+          <div className="flex grow flex-col gap-2">
+            <div className="overflow-hidden text-ellipsis whitespace-nowrap">{item.name}</div>
+            <AccountsLogoStack addresses={addresses} />
+          </div>
+          {onCollapse && (
+            <Action onClick={onCollapse}>
+              <ChevronDownIcon
+                className={classNames("transition-transform", collapsed && "-rotate-90")}
+              />
+            </Action>
+          )}
+          {collapsed ? (
+            <MoreHorizontalIcon className="shrink-0" />
+          ) : (
+            <ChevronDownIcon className="text-body-disabled shrink-0" />
+          )}
+          {onDelete && (
+            <Action onClick={onDelete}>
+              <TrashIcon />
+            </Action>
+          )}
+          {clone && childCount && childCount > 1 ? (
+            <span
+              className={classNames(
+                "bg-black-tertiary text-primary rounded-xs absolute -right-5 -top-5 flex h-12 w-12 items-center justify-center text-xs",
+                clone && "select-none"
+              )}
+            >
+              {childCount}
+            </span>
+          ) : null}
+        </div>
+      </TreeItemWrapper>
+    )
+  }
+)
+TreeFolderItem.displayName = "TreeFolderItem"
+
+const TreeItemWrapper = forwardRef<HTMLLIElement, Props & { children?: ReactNode | undefined }>(
+  (
+    { children, clone, depth, disableInteraction, disableSelection, ghost, indentationWidth },
+    ref
+  ) => (
+    <li
+      className={classNames(
+        "-mb-[1px] list-none",
+        clone && "pointer-events-none",
+        ghost && "opacity-50",
+        disableSelection && "select-none",
+        disableInteraction && "pointer-events-none"
+      )}
+      ref={ref}
+      style={{ paddingLeft: clone ? undefined : `${indentationWidth * depth}px` }}
+    >
+      {children}
+    </li>
+  )
+)
+TreeItemWrapper.displayName = "TreeItemWrapper"
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const DragButton = (props: any) => (
+  <DragIcon className="text-grey-750 -mx-4 shrink-0 text-xl" {...props} />
 )
